requirements.txt
python-jobspy
requests
python-dotenv
fastapi
uvicorn[standard]
playwright
jinja2
python-multipart
